+++ src/Main.tsx
@@ -1,16 +1,16 @@
-import React, { FC } from 'react';
-import { NavigationContainer } from '@react-navigation/native';
+import React, { FC } from "react";
+import { NavigationContainer } from "@react-navigation/native";
 
-import BottomTabs from './components/BottomNavigation';
+import BottomTabs from "./navigation/BottomNavigation";
 
-interface IMainProps { }
+interface IMainProps {}
 
-const Main: FC<IMainProps> = props => {
-    return (
-        <NavigationContainer>
-            <BottomTabs />
-        </NavigationContainer>
-    );
+const Main: FC<IMainProps> = (props) => {
+  return (
+    <NavigationContainer>
+      <BottomTabs />
+    </NavigationContainer>
+  );
 };
 
 export default Main;