+++ src/constans/Layout.ts
@@ -1,8 +1,8 @@
-import { Dimensions } from 'react-native';
-import Constants from 'expo-constants';
+import { Dimensions } from "react-native";
+import Constants from "expo-constants";
 
-const width = Dimensions.get('window').width;
-const height = Dimensions.get('window').height;
+const width = Dimensions.get("window").width;
+const height = Dimensions.get("window").height;
 
 const wp = (percentage: number) => {
   const value = (percentage * width) / 100;
@@ -12,9 +12,9 @@ const wp = (percentage: number) => {
 export default {
   window: {
     width,
-    height
+    height,
   },
   wp,
   statusBar: Constants.statusBarHeight,
-  isSmallDevice: width < 375
+  isSmallDevice: width < 375,
 };