+++ src/reducers/example.ts
@@ -1,23 +1,23 @@
-import * as actionTypes from '../actions/types/example';
+import * as actionTypes from "../actions/types/example";
 
 export interface IExampleReducer {
-    exampleState: string;
+  exampleState: string;
 }
 
 const defaultState = (): IExampleReducer => ({
-    exampleState: undefined
+  exampleState: undefined,
 });
 
 export default (state = defaultState(), action: any): IExampleReducer => {
-    switch (action.type) {
-        case actionTypes.EXAMPLE_ACTIONS: {
-            return {
-                ...state,
-                exampleState: action.text
-            };
-        }
-        default: {
-            return state;
-        }
+  switch (action.type) {
+    case actionTypes.EXAMPLE_ACTIONS: {
+      return {
+        ...state,
+        exampleState: action.text,
+      };
     }
+    default: {
+      return state;
+    }
+  }
 };